+++ packages/article-flag/src/article-flag-prop-types.js
@@ -1,0 +1,12 @@
+import PropTypes from "prop-types";
+import { colours } from "@times-components/styleguide";
+
+const articleFlagPropTypes = {
+  color: PropTypes.string
+};
+
+const articleFlagDefaultProps = {
+  color: colours.functional.primary
+};
+
+export { articleFlagPropTypes, articleFlagDefaultProps };