+++ packages/article-in-depth/src/article-meta/article-meta-prop-types.js
@@ -1,0 +1,9 @@
+import PropTypes from "prop-types";
+
+const metaPropTypes = {
+  byline: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
+  publicationName: PropTypes.string.isRequired,
+  publishedTime: PropTypes.string.isRequired
+};
+
+export default metaPropTypes;