+++ packages/article-in-depth/src/styles/index.js
@@ -1,0 +1,16 @@
+import { StyleSheet } from "react-native";
+import sharedStyles from "./shared";
+
+const nativeStyles = {
+  ...sharedStyles,
+  standFirst: {
+    ...sharedStyles.standFirst,
+    lineHeight: 25
+  }
+};
+
+const styles = StyleSheet.create({
+  ...nativeStyles
+});
+
+export default styles;