+++ packages/article/src/article.js
@@ -1,10 +1,12 @@
 import React from "react";
 import ArticleMagazineComment from "@times-components/article-magazine-comment";
+import ArticleInDepth from "@times-components/article-in-depth";
 import ArticleMagazineStandard from "@times-components/article-magazine-standard";
 import ArticleMainStandard from "@times-components/article-main-standard";
 import ArticleMainComment from "@times-components/article-main-comment";
 
 export const templates = {
+  indepth: ArticleInDepth,
   magazinecomment: ArticleMagazineComment,
   magazinestandard: ArticleMagazineStandard,
   maincomment: ArticleMainComment,