+++ packages/ad/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.10](https://github.com/newsuk/times-components/compare/@times-components/ad@2.2.9...@times-components/ad@2.2.10) (2018-12-17)
+
+
+### Bug Fixes
+
+* correct storybook adslot config ([#1579](https://github.com/newsuk/times-components/issues/1579)) ([53dc125](https://github.com/newsuk/times-components/commit/53dc125))
+
+
+
+
+
 ## [2.2.9](https://github.com/newsuk/times-components/compare/@times-components/ad@2.2.8...@times-components/ad@2.2.9) (2018-12-13)
 
 **Note:** Version bump only for package @times-components/ad