+++ packages/article-magazine-standard/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.6](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@0.2.5...@times-components/article-magazine-standard@0.2.6) (2018-12-17)
+
+**Note:** Version bump only for package @times-components/article-magazine-standard
+
+
+
+
+
 ## [0.2.5](https://github.com/newsuk/times-components/compare/@times-components/article-magazine-standard@0.2.4...@times-components/article-magazine-standard@0.2.5) (2018-12-14)
 
 **Note:** Version bump only for package @times-components/article-magazine-standard